requirements.txt doesn't call setup.py
The -e option invokes 'python setup.py install .', which in travis tries
install, but fails, because the binding isn't built yet. So we explicitly list
the dependencies in the requirements.txt instead. It would be nice if there was
some way to extract only the dependencies from setup.py, and not fail if the
binding wasn't actually there yet. Probably there is, I am just lazy for now
:-)