remove bytes_to_cdata
This function is Very Dangerous due to the cffi semantics. Because the name the
memory is initially assigned to is deleted before any user of the function ever
sees the memory, every call has the potential to free the memory before it is
used. See https://groups.google.com/forum/#!topic/python-cffi/Bk4CniFZUFY for a
complete explanation.