Change license to Apache 2.0
There are several reasons to do this, namely because it is similar in spirit to
MIT license, but it is What Everyone Else uses. The reason everyone else uses
it is because it contains a patent grant along with a grant to the rights of
the software. This doesn't apply in xcffib's case (I have no patents), but
there is existing infrastructure in place to understand and use Apache 2.0
licenses, where there is not for the pseudo-ambiguous MIT/Expat license (e.g.
Debian distributes a copy of Apache-2.0, but distributes no copy of any MIT
license).

Closes #30