Differentiate between cffi unpackers and memory unpackers
In order to create unions synthetically, the best API I can come up with is to
make people encode their data back to binary form, and let us unpack it as a
union (vs. making people specify it multiple times in various encodings). So,
we need the concept of a MemoryUnpacker for people who are interested in making
their own data in memory.