Add a demo of how to pack() using the broken pack
The mangling of the result of the pack() buffer here is quite annoying, but at
least this is more useful than what was in xpyb. I am still thinking about how
best to design an API to pack messages nicely, but at least this shows that it
is possible to do now.