Correctly render enums with explicit values in the middle
It looks like someone at some point between 1.7 and 1.10 made all of these
explicit (i.e. they have an explicit value parameter for each value, not just
some of them). However, it is still a bug in the places where people didn't do
this, and also if people are trying to use xcffibgen on xcb-protos < 1.10.