Add cffi compatibility layer for PyPy < 2.6
Allow for cffi < 1.0 to be used by PyPy, since PyPy ships with a built
in cffi module (needed for proper JIT'ing) which can't be overwritten by
installing a newer version, and the current version of PyPy (2.5.1) has
cffi 0.9.2, this puts back some of the hacks to get cffi working for
older versions of cffi if and only if running with PyPy.

This can be reverted when PyPy 2.6 is released.