Use ABI-based cffi
Link against cffi libraries using dlopen rather than compiling against
the xcb library.

While this may complicate things, we need to maintain the values of
defines and it may not be immediately apparent on install that there
are problems with the libxcb install, this allows us to build xcb
support into carocffi using the new cffi-1.0 features.