Unpack Switch'es
Unpack Switch XCB objects.  This is a bit convoluted, since each case in
the switch may(?) contain a set of Statements, but we'll assume we only
have one statement.

The structElemToPyUnpack is expanded to have as a return parameter a
list of 3 elements, the first being the expression, the second being the
padding information, and last one being the cases for the switch.  The
previous behavior is passing a singleton list with the first and second
elements as the expressions previously used for these and Nothing for
the cases.

Some assumputions are made about nesting of switches and what data types
are set by the cases, this could do with some refactoring, but this
accomplishes the task with minimal re-working of functions.  This will
be a problem if we ever try to do xkb (nested switches, etc).