atoms are always little endian
It seems that atom values are always little endian. In principle this
should be specified somewhere, but I could only find a mention of something
related here:

https://groups.google.com/forum/#!topic/golang-nuts/mJSez6b5pHM

Anyway, as-such, this patch contains two fixes:

1. don't use struct.pack for packing WM_TAKE_FOCUS in the test. On little
   endian architectures this is a no-op, and on big-endian ones, it is
   actually incorrect, since it changes the order to native order instead
   of leaving it little endian, the byte order was that we got back from X.
2. to_atoms should always encode things as little endian.

I reviewed our other uses of "=" in struct.pack, and I think they are all
correct.

Closes #77

Signed-off-by: Tycho Andersen <tycho@tycho.ws>