travis: install xcb-types after installing deps
Recent versions of cabal seem to download and install the packaged version
of a package when doing cabal install --only-dependencies, even if you have
a locally installed one that you built by hand. So, let's build xcb-types
after we install xcffib's dependencies. This means slightly more work for
travis because it'll have to build xcb-types twice, but it's a small
package and it doesn't work the other way, so here we are :)

Closes #88

Signed-off-by: Tycho Andersen <tycho@tycho.ws>