connection: don't automatically disconnect() on connection error
It turns out that qtile has some tests that depend on this behavior (i.e.
that once you get one ConnectionException(), you will keep getting them),
so this convenience of automatically disconnecting won't work.

Let's just drop it, but keep the other half of eb856bb199c6 ("connection:
change the semantics for invalid()+disconnect()") that drops the invalid()
call from disconnect(), since that was really the meat of the fix.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>