tests: wait for response from ChangeWindowattributes()
every once in a while, we see an indefinite hang in the tests. nobody has
ever complained about anything like that in the actual binding, so it
stands to reason the bug is actually in the tests.

my best guess is the following:

1. ChangeWindowAttributes() to get map/create/destroy events
2. spawn xeyes
3. flush the connection

If xeyes happened to spawn and its events triggered before the connection
got flushed and the monitoring of events actually enabled, we could
conceivably miss these events and hang forever. instead, let's use a
checked request for the ChangeWindowAttributes() part, and check the
result. then we don't need these heavy-weight flushes (that were in the
wrong order anyway... at all)

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>