ci: use "latest" version of haskell packages
We mostly don't worry about building the haskell code with various
compilers, as it doesn't change very much any more. As such, it's annoying
to keep having these version bump PRs for haskell versions. Let's just
build with "latest", and hopefully there won't be too much build breakage.
Of course we can revert to setting specific versions if it turns out there
is... but I sort of doubt it.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>