union unpack: don't use partitionEithers
This is similar to the previous commit where we can't re-order fields in a
union based on whether they're a base or complex type. Obviously this
implementation sucks, but it's not super obvious what to do here, and
this case currently isn't hit in xcb-proto anyway. Rather than implement
something that's maybe correct and unused, let's just leave it as an error
until someone needs it. At least now it won't fail silently, though.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>