generator: handle paramrefs appropriately
paramref is really another hack for dynamically sized types (everywhere
else in the binding would declare these things as a list...).

We get:

    xcffib/xinput.py:288:54: F821 undefined name 'num_axes'

This object doesn't actually know the number of axes it has on the wire..
those come from a different request, and we're supposed to grab it from the
context in which the request was made.

Luckily, there is only one of these, so we can just hardcode the behavior
and explicitly error out if anyone ever adds another one.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>