do not depend on cffi for PyPy
Make the dependency on `cffi` conditional to non-PyPy interpreters,
in order to make the package installable on modern PyPy versions
and ensure that the wheel metadata is static, irrespectively of which
Python implementation is used to build it.  The previous hack would
attempt to install exactly the same `cffi` version that was used on PyPy
which in the best case was redundant, and in the worst case made
the package non-installable if PyPy used a non-public `cffi` version.

Fixes #166