generator: add support for xprint
sometime between when I first wrote this readme and now, someone (me?
someone else?) must have hacked together enough support for xprint such
that it parses correctly now. So let's emit it.

Obviously this is entirely untested, but it's Probably Right, given the
amount of inspection the rest of the binding has seen. Famous last words :)

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>