Revert "cffi api mode"
This reverts commit fb83a360a491b0a5966c7f0f0df94b6f47421fb0.

This fanciness unfortunately does not work transparently to library
consumers like cairocffi, so they will need some patches to understand this
before we can release with it. This "fixes" github issue #179, but I'm
going to leave it open so we can have the discussion about how to fix it
there.