ci: don't use pytest-xdist
This should really have been part of f1fcbb5f4043 ("ci: don't do things in
parallel"), we should also disable pytest-xdist.

We are still getting the stack trace/xserver error messages mentioned in
the above commit r.e. socket already bound, etc. So it seems as though the
find_display() locking logic is not quite bulletproof.

I have tested it alone and been unable to reproduce this failure, so let's
just run these tests serially for now. My suspicion is that there is
something with the pytest harness that's forcing a fork or closing the
files, etc.

One option would be to experiment with locking stuff completely outside of
x11's lock mechanisms, but that also feels a bit ugly. For now, we have
this...

(Maybe we should delete all this locking logic instead of pretending that
it works? But it's likely the fix is very simple, so...)

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>