fix __repr__ for structs
c7d7bf83b522 ("add a better __repr__") seems to not really have worked; I'm
not sure why I thought it did, but the results are bogus (i.e. there is no
members in the resulting structures).

Let's just define our own on Protobj, without @dataclass hacks. it's not
that much code: we just skip a few irrelevant things, and look in our own
__dir__.

We keep the list special casing, and add tests for all of these, and
hashing, which was something I was using that broke.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>