switch to pyproject.toml
This is the "way forward" for packaging, so let's use it.

The one thing we lose is the nice error message if you haven't built the
binding locally and you try to install. I expect that these days the only
person who really builds the binding locally is me, and I do it via the
makefile, so hopefully we don't really need that error message any more.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>