don't use __getattr__for Checked reasoning
While profiling Qtile, I noticed:

    2   xcffib/__init__.py:401                   17.3 KiB
        real = name[: -len("Checked")]

we generate a *ton* of memory allocations for these Checked and Unchecked
calls where the default is otherwise. I like this flexibility, but it is
annoying that it generates so many allocations.

So instead, let's generate a bigger vtable :). We can generate Checked and
Unchecked variants of each method, that call the underlying version with
the boolean flag. This way there are no (runtime) allocations, "just" a
vtable that's twice as big. But e.g. for xproto, it was already huge
anyway, so I don't expect a huge performance hit here...

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>